[FEAT] Dashboard Data Ingestion
This commit adds the ability to
ingest data from the `sites` and
`advertisements` app into the
`dashboard` app.

The `dashboard` app is a separate
app that is used to store data
that is used for the `DenormalizedAdvertisement`
model.

A management command called `ingest_data`
has been added to the `dashboard` app
that will ingest data from the `sites` and
`advertisements` app into the `dashboard`
app.

`models.py` in `dashboard` has been modified
to add the `DenormalizedAdvertisement` model.

`routers.py` in `dashboard` has been modified
to add the `DashboardRouter` class.

`migrations/0001_initial.py` in `dashboard`
has been modified to add the `DenormalizedAdvertisement`
model.